removed button and added jest config for svg test not passing
