+++ packages/ui-components/__tests__/button.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {fireEvent, render, screen} from '@testing-library/react';
 
-import {Button} from '../src/components/button';
+import {ButtonBase as Button} from '../src/components/button/buttonBase';
 
 describe('Button', () => {
   function setup(args: any) {
