+++ packages/ui-components/jest.config.js
@@ -1,0 +1,6 @@
+const {createJestConfig} = require('tsdx/dist/createJestConfig');
+const {paths} = require('tsdx/dist/constants');
+
+const config = createJestConfig(undefined, paths.appRoot);
+config.transform['^.+\\.svg$'] = 'jest-svg-transformer';
+module.exports = config;
