+++ packages/ui-components/src/components/button/index.tsx
@@ -1,4 +1,3 @@
-export {ButtonProps, Button} from './button';
 export * from './buttonText';
-export {WalletButton, WalletButtonProps} from './walletButton';
 export * from './buttonIcon';
+export {WalletButton, WalletButtonProps} from './walletButton';
