+++ packages/ui-components/src/components/button/walletButton.tsx
@@ -1,11 +1,30 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import {SizedButton} from './button';
 import {Avatar} from '../avatar';
 import {Spinner} from '../spinner';
 import {shortenAddress} from '../../utils/addresses';
 import {IconPerson} from '../icons';
+import type {ButtonBaseProps} from './buttonBase';
+
+// TODO: refactor to remove sized button to baseButton
+
+type SizedButtonProps = {
+  size: ButtonBaseProps['size'];
+};
+
+/**
+ * Extends the html button element with the desired size.
+ * Furthermore, the button comes with rounded corners and a focus ring, as this
+ * is shared by all buttons in this library.
+ */
+export const SizedButton = styled.button.attrs(({size}: SizedButtonProps) => {
+  let baseClasses =
+    'px-2 focus:outline-none focus:ring-2 focus:ring-primary-500 font-bold';
+  let sizedClasses =
+    size === 'medium' ? 'py-1.5 rounded-2xl' : 'py-1 rounded-xl text-sm';
+  return {className: `${baseClasses} ${sizedClasses}`};
+})<SizedButtonProps>``;
 
 export type WalletButtonProps = {
   /**
@@ -55,10 +74,12 @@ export const WalletButton = ({
   };
 
   return (
-    <StyledButton size={'small'} isSelected={isSelected} {...props}>
-      <StyledLabel {...{isLoading}}>{shortenAddress(label)}</StyledLabel>
-      {LoadAvatar()}
-    </StyledButton>
+    <>
+      <StyledButton size={'small'} isSelected={isSelected} {...props}>
+        <StyledLabel {...{isLoading}}>{shortenAddress(label)}</StyledLabel>
+        {LoadAvatar()}
+      </StyledButton>
+    </>
   );
 };
