+++ packages/ui-components/stories/actionListItem.stories.tsx
@@ -27,7 +27,6 @@ export const TitleOnly = Template.bind({});
 TitleOnly.args = {
   title: 'Action List Item',
   icon: <IconTurnOff />,
-  wide: true,
 };
 
 export const Disabled = Template.bind({});
@@ -35,13 +34,5 @@ Disabled.args = {
   title: 'Action List Item',
   subtitle: 'Action List Item Subtitle',
   disabled: true,
-  icon: <IconChevronRight />,
-};
-
-export const Wide = Template.bind({});
-Wide.args = {
-  title: 'Action List Item',
-  subtitle: 'Action List Item Subtitle',
-  wide: true,
   icon: <IconChevronRight />,
 };
