+++ packages/ui-components/stories/daoCard.stories.tsx
@@ -14,9 +14,7 @@ Default.args = {
   daoName: 'DaoName',
   wide: false,
   daoAddress: '0x6720000000000000000000000000000000007739',
-  src:
-    'https://banner2.cleanpng.com/20180325/sxw/kisspng-computer-icons-avatar-avatar-5ab7529a8e4e14.9936310115219636745829.jpg',
-  switchLabel: 'Switch',
+  src: 'https://banner2.cleanpng.com/20180325/sxw/kisspng-computer-icons-avatar-avatar-5ab7529a8e4e14.9936310115219636745829.jpg',
   onClick: () => {
     alert('Pressing this button would allow to change DAO.');
   },
@@ -28,9 +26,7 @@ WithoutSwitch.args = {
   wide: true,
   includeSwitch: false,
   daoAddress: '0x6720000000000000000000000000000000007739',
-  src:
-    'https://banner2.cleanpng.com/20180325/sxw/kisspng-computer-icons-avatar-avatar-5ab7529a8e4e14.9936310115219636745829.jpg',
-  switchLabel: 'Switch',
+  src: 'https://banner2.cleanpng.com/20180325/sxw/kisspng-computer-icons-avatar-avatar-5ab7529a8e4e14.9936310115219636745829.jpg',
   onClick: () => {
     alert('Pressing this button would allow to change DAO.');
   },
