+++ packages/ui-components/stories/popover.stories.tsx
@@ -1,6 +1,6 @@
 import React, {ReactNode} from 'react';
 import {Meta, Story} from '@storybook/react';
-import {Button, Popover, PopoverProps} from '../src';
+import {ButtonText, Popover, PopoverProps} from '../src';
 
 export default {
   title: 'Components/Popover',
@@ -14,7 +14,7 @@ const Template: Story<PopoverProps> = args => (
   <Popover {...args}>{args.children}</Popover>
 );
 
-const trigger: ReactNode = <Button label="Trigger" />;
+const trigger: ReactNode = <ButtonText label="Trigger" />;
 const popoverContent: ReactNode = (
   <div>
     Lorem ipsum dolor sit amet consectetur adipisicing elit. Maxime mollitia,
